refactor: extract dialogue prompt assembly
